requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
distro
exceptiongroup
fastapi
filelock
fsspec
groq
h11
httpcore
httpx
huggingface-hub
idna
Jinja2
jiter
joblib
jsonpatch
jsonpointer
langchain-core
langchain-openai
langsmith
MarkupSafe
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
openai
orjson
packaging
pillow
pinecone
pinecone-client
pinecone-plugin-interface
pydantic
pydantic_core
PyPDF2
python-dateutil
python-dotenv
python-multipart
PyYAML
regex
requests
requests-toolbelt
safetensors
scikit-learn
scipy
sentence-transformers
six
sniffio
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
triton
typing-inspection
typing_extensions
urllib3
uvicorn
websockets
zstandard